Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
bcrypt
cryptography